+++ src/components/TodoList.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import TodoItem from './TodoItem';
+
+function TodoList(props) {
+    const {title, items} = props;
+
+    return (
+        <div className="todolist">
+            <h1>{title.toUpperCase()}</h1>
+            <ul className="list-unstyled">
+                {items.map(item => <TodoItem key={item.id} data={item}/>)}
+            </ul>
+        </div>
+    );
+}
+
+export default TodoList;
