+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './App';
+import App from './components/App';
 // Added bootstrap
 import 'bootstrap/dist/css/bootstrap.css';
 import './custom.scss';
